move RunView() to trace.py
