+++ aim/sdk/trace.py
@@ -12,12 +12,15 @@ from aim.storage.context import Context
 from aim.storage.hashing import hash_auto
 from aim.storage.proxy import AimObjectProxy
 from aim.storage.query import RestrictedPythonQuery
+from aim.storage.structured.entities import StructuredObject
+from aim.storage.treeview import TreeView
 
 from typing import Any, Generic, Iterator, List, TypeVar, Union
 from typing import TYPE_CHECKING
 
+
 if TYPE_CHECKING:
-    from aim.sdk.run import Run, RunView
+    from aim.sdk.run import Run
     from aim.sdk.repo import Repo
     from pandas import DataFrame
 
@@ -164,6 +167,35 @@ class Trace(Generic[T]):
         return df
 
 
+class RunView:
+
+    def __init__(self, run: 'Run'):
+        self.db = run.repo.structured_db
+        self.hashname = run.hashname
+        self.structured_run_cls: type(StructuredObject) = self.db.run_cls()
+        self.meta_run_tree: TreeView = run.meta_run_tree
+        self.meta_run_attrs_tree: TreeView = run.meta_run_attrs_tree
+
+    def __getattr__(self, item):
+        if item in self.structured_run_cls.fields():
+            return getattr(self.db.caches['runs_cache'][self.hashname], item)
+        else:
+            return self[item]
+
+    def __getitem__(self, key):
+        return AimObjectProxy(lambda: self.meta_run_attrs_tree.collect(key), view=self.meta_run_attrs_tree.view(key))
+
+    def get(
+        self,
+        key,
+        default: Any = None
+    ) -> AimObject:
+        try:
+            return self.__getitem__(key)
+        except KeyError:
+            return default
+
+
 class ContextView:
     def __init__(self, context: dict):
         self.context = context
@@ -189,7 +221,7 @@ class ContextView:
 
 
 class MetricView:
-    def __init__(self, name: str, context: dict, run_view: 'RunView'):
+    def __init__(self, name: str, context: dict, run_view: RunView):
         self.name = name
         self.run = run_view
         self._context = context
@@ -276,7 +308,7 @@ class RunTraceCollection(TraceCollection):
             if not self.query_traces:
                 statement = True
             else:
-                run_view = run.view()
+                run_view = RunView(run)
                 metric_view = MetricView(metric_name, ctx.to_dict(), run_view)
                 statement = self.query_traces.match(
                     run=run_view,
@@ -328,7 +360,7 @@ class QueryRunTraceCollection(TraceCollection):
             if not self.query:
                 statement = True
             else:
-                run_view = run.view()
+                run_view = RunView(run)
                 statement = self._query.match(run=run_view)
             if not statement:
                 continue
